Compass -> Binnacle in test suite
